+++ adblock-android-webview/src/androidTest/kotlin/org/adblockplus/libadblockplus/android/webview/test/AdBlockingWhitelistingTest.kt
@@ -24,6 +24,7 @@ import com.github.tomakehurst.wiremock.client.WireMock.urlPathEqualTo
 import org.adblockplus.libadblockplus.android.AndroidHttpClient
 import org.adblockplus.libadblockplus.android.Utils
 import org.adblockplus.libadblockplus.android.webview.AdblockWebView
+import org.adblockplus.libadblockplus.android.webview.BaseSiteKeyExtractor
 import org.adblockplus.libadblockplus.android.webview.SiteKeyHelper
 import org.adblockplus.libadblockplus.android.webview.imageIsBlocked
 import org.adblockplus.libadblockplus.android.webview.imageIsNotBlocked
@@ -34,8 +35,8 @@ import org.adblockplus.libadblockplus.security.SlowSignatureVerifierWrapper
 import org.adblockplus.libadblockplus.sitekey.PublicKeyHolderImpl
 import org.adblockplus.libadblockplus.sitekey.SiteKeysConfiguration
 import org.junit.Assert.assertNotNull
+import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
-import org.junit.Ignore
 import org.junit.Test
 import timber.log.Timber
 import wiremock.org.apache.http.HttpStatus
@@ -314,13 +315,14 @@ class AdBlockingWhitelistingTest : BaseAdblockWebViewTest() {
     }
 
     @Test
-    @Ignore("Ignored until we have DP-1581 fixed")
-    fun testWhitelistedMainFrameWithSitekey() {
+    fun testWhitelistedMainFrameResourcesHeldUntilSiteKey() {
         val blockingRule = greenImage
 
-        // 30 seconds delay should be enough to force all the other resources requests coming
-        // to simulate "resource loading check happens before sitekey extracted"
-        val (sitekey, signature) = initSitekey(indexPageUrl, 30_000)
+        // A half of max delay ensures that resources are held and whitelisted after the sitekey
+        // verification concludes.
+        // If the resources are not held the default blocking rule will apply.
+        val (sitekey, signature) = initSitekey(indexPageUrl,
+                (BaseSiteKeyExtractor.RESOURCE_HOLD_MAX_TIME_MS / 2).toLong())
 
         // whitelist main frame with sitekey
         val whitelistingRule = "@@\$subdocument,document,sitekey=$sitekey"
@@ -344,5 +346,43 @@ class AdBlockingWhitelistingTest : BaseAdblockWebViewTest() {
         onAdblockWebView()
             .check(imageIsNotBlocked(blockedImageWithSlashId))
             .check(elementIsNotElemhidden(blockedImageWithSlashId))
+    }
+
+    @Test
+    fun testSkipWhitelistingWhenSitekeyVerificationTimeouts() {
+        val blockingRule = greenImage
+
+        // Twice the max delay is enough to simulate the timeout of holding of the resource loading
+        // before sitekey is extracted. The default rule is applied in this case.
+        val (sitekey, signature) = initSitekey(indexPageUrl,
+                (BaseSiteKeyExtractor.RESOURCE_HOLD_MAX_TIME_MS * 2).toLong())
+
+        // whitelist main frame with sitekey
+        val whitelistingRule = "@@\$subdocument,document,sitekey=$sitekey"
+
+        val (blockedResources, whitelistedResources) = load(
+            listOf(blockingRule, whitelistingRule),
+            """
+            |<html data-adblockkey="$signature">
+            |<body>
+            |  <img id="$blockedImageWithSlashId" src="/$greenImage"/>
+            |</body>
+            |</html>
+            |""".trimMargin())
+
+        // main frame resource is not whitelisted
+        assertNull(whitelistedResources.find {
+            it.requestUrl == "${wireMockRule.baseUrl()}/$greenImage" && it.parentFrameUrls.size == 1
+        })
+
+        // main frame resource is blocked
+        assertNotNull(blockedResources.find {
+            it.requestUrl == "${wireMockRule.baseUrl()}/$greenImage" && it.parentFrameUrls.size == 1
+        })
+
+        // green image in main frame is blocked with the default blocking rule
+        onAdblockWebView()
+            .check(imageIsBlocked(blockedImageWithSlashId))
+            .check(elementIsElemhidden(blockedImageWithSlashId))
     }
 }