+++ adblock-android-webview/src/main/java/org/adblockplus/libadblockplus/android/webview/SiteKeyExtractor.java
@@ -64,6 +64,23 @@ public interface SiteKeyExtractor
                                             WebResourceRequest request);
 
   /**
+   * Notifies about starting of a new page
+   */
+  void startNewPage();
+
+  /**
+   * Blocks the calling thread while checking the sitekey
+   * <p>
+   * Will be removed later in a favor of setting internal WebViewClient
+   * for every SiteKeyExtractor
+   *
+   * @param request from the
+   *                {@link android.webkit.WebViewClient#shouldInterceptRequest(WebView, WebResourceRequest)}
+   * @return true if had to wait
+   */
+  boolean waitForSitekeyCheck(final WebResourceRequest request);
+
+  /**
    * This method is called by the {@link AdblockWebView} during
    * {@link AdblockWebView#setSiteKeysConfiguration(SiteKeysConfiguration)}
    * <p/>