+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Decision on blocking the resources is now held until the sitekey is checked - [!355](https://gitlab.com/eyeo/adblockplus/libadblockplus-android/-/merge_requests/355) 
+
 ### Added
 - Semantic Versioning is introduced. [!358](https://gitlab.com/eyeo/adblockplus/libadblockplus-android/-/merge_requests/358)